EX-2312: add tickSpacing to extra when init univ3 pools

+++ pkg/source/solidly-v3/type.go
@@ -58,7 +58,7 @@ type Tick struct {
 type Extra struct {
 	Liquidity    *big.Int `json:"liquidity"`
 	SqrtPriceX96 *big.Int `json:"sqrtPriceX96"`
-	TickSpacing  int32    `json:"tickSpacing"`
+	TickSpacing  uint64   `json:"tickSpacing"`
 	Tick         *big.Int `json:"tick"`
 	Ticks        []Tick   `json:"ticks"`
 }
